Extracted DeviceMetaData from MieleBridgeHandler.
